# Modest speed-ups for the _sum() helper function in statistics module