Three small speed-ups for our slowest helper method:

1) Replace the ``dict.partials.get`` call with ``defaultdict(int)``.  On bigger datasets, this is about a 25% win. On smaller datasets, it is a little slower that the ``get()`` approach. The reason is that *get* loads the constant zero regardless of whether denominator has been seen before.  In contrast, the *defaultdict* constructs a zero with the slow call to ``int()`` but only does so when the denominator has not been seen before.  So, datasets with many repeated denominators avoid this cost.

2) Now that *int*, *float*, *Decimal*, and *Fraction* all support ``as_integer_ratio()``, it makes sense to inline that common case from ``exact_ratio()``.  This avoids function call overhead for every datum.

3) Eliminate overhead of repeated calls to ``Fraction.__new__`` and ``Fraction.__add__`` by summing directly to the least common multiple using fast inline integer arithmetic.  The final *Fraction()* instance reduces the fraction if needed.

**Baseline timings**

```
# baseline: small run with 10 elements
$ python -m timeit -r 11 -s 'from random import expovariate' -s 'from statistics import _sum' -s 'data=[expovariate(1.0) for i in range(10)]' '_sum(data)'
10000 loops, best of 11: 28.4 usec per loop

# baseline:  medium run with 100 elements
$ python -m timeit -r 11 -s 'from random import expovariate' -s 'from statistics import _sum' -s 'data=[expovariate(1.0) for i in range(100)]' '_sum(data)'
5000 loops, best of 11: 61.6 usec per loop

#baseline: long run with 10,000 elements
$ python -m timeit -r 11 -s 'from random import expovariate' -s 'from statistics import _sum' -s 'data=[expovariate(1.0) for i in range(10_000)]' '_sum(data)'
50 loops, best of 11: 5.28 msec per loop
```

**Patched timings**

```
# patched: small run with 10 elements
$ python -m timeit -r 11 -s 'from random import expovariate' -s 'from statistics import _sum' -s 'data=[expovariate(1.0) for i in range(10)]' '_sum(data)'  
20000 loops, best of 11: 9.97 usec per loop

# patched: medium run with 100 elements
$ python -m timeit -r 11 -s 'from random import expovariate' -s 'from statistics import _sum' -s 'data=[expovariate(1.0) for i in range(100)]' '_sum(data)'   
5000 loops, best of 11: 54.3 usec per loop

# patched: long run with 10,000 elements
$ python -m timeit -r 11 -s 'from random import expovariate' -s 'from statistics import _sum' -s 'data=[expovariate(1.0) for i in range(10_000)]' '_sum(data)'
50 loops, best of 11: 4.74 msec per loop
```