@rhettinger 2021-09-21
Am going to do more research to see if the improvements can be made more uniform (this PR makes some cases worse).